[refactor] Data source: merge GoogleMapsApIClient into GoogleApiDataSource
