+++ composeApp/src/commonMain/kotlin/uk/ryanwong/gmap2ics/data/datasources/googleapi/GetPlaceDetailsAPIErrorException.kt
@@ -1,0 +1,9 @@
+/*
+ * Copyright (c) 2024. Ryan Wong (hello@ryanwebmail.com)
+ */
+
+package uk.ryanwong.gmap2ics.data.datasources.googleapi
+
+class GetPlaceDetailsAPIErrorException(apiErrorMessage: String) : Exception() {
+    override val message = "⛔️ Error getting API results: $apiErrorMessage"
+}
