+++ composeApp/src/commonTest/kotlin/uk/ryanwong/gmap2ics/data/datasources/googleapi/KtorGoogleApiDataSourceTestData.kt
@@ -2,11 +2,15 @@
  * Copyright (c) 2022-2024. Ryan Wong (hello@ryanwebmail.com)
  */
 
-package uk.ryanwong.gmap2ics.data.datasources.googleapi.ktor
+package uk.ryanwong.gmap2ics.data.datasources.googleapi
 
-internal object GoogleMapsApiClientImplTestData {
+import uk.ryanwong.gmap2ics.data.models.places.GeometryDto
+import uk.ryanwong.gmap2ics.data.models.places.LocationDto
+import uk.ryanwong.gmap2ics.domain.models.timeline.PlaceDetails
 
-    const val PLACE_DETAILS_GREG_AVE = """{
+internal object KtorGoogleApiDataSourceTestData {
+
+    const val PLACE_DETAILS_GREG_AVE_JSON = """{
    "html_attributions" : [],
    "result" : {
       "address_components" : [
@@ -82,4 +86,20 @@ internal object GoogleMapsApiClientImplTestData {
    },
    "status" : "OK"
 }"""
+
+    val PLACE_DETAILS_GREG_AVE_DOMAIN_MODEL = PlaceDetails(
+        formattedAddress = "8 Greg Ave, Bollington, Macclesfield SK10 5HR, UK",
+        formattedPhoneNumber = null,
+        geometry = GeometryDto(locationDto = LocationDto(lat = 53.2945761, lng = -2.114387)),
+        icon = "https://maps.gstatic.com/mapfiles/place_api/icons/v1/png_71/geocode-71.png",
+        name = "8 Greg Ave",
+        placeId = "ChIJ43NG7NxLekgR7NFDJzb-WXw",
+        reference = "ChIJ43NG7NxLekgR7NFDJzb-WXw",
+        types = listOf("premise"),
+        url = "https://maps.google.com/?q=8+Greg+Ave,+Bollington,+Macclesfield+SK10+5HR,+UK&ftid=0x487a4bdcec4673e3:0x7c59fe362743d1ec",
+        userRatingsTotal = null,
+        utcOffset = 60,
+        vicinity = "Bollington",
+        website = null,
+    )
 }
