+++ composeApp/src/commonTest/kotlin/uk/ryanwong/gmap2ics/data/datasources/local/fakes/FakeGoogleApiDataSource.kt
@@ -4,7 +4,7 @@
 
 package uk.ryanwong.gmap2ics.data.datasources.local.fakes
 
-import uk.ryanwong.gmap2ics.data.datasources.googleapi.GoogleApiDataSource
+import uk.ryanwong.gmap2ics.data.datasources.googleapi.interfaces.GoogleApiDataSource
 import uk.ryanwong.gmap2ics.domain.models.timeline.PlaceDetails
 
 class FakeGoogleApiDataSource : GoogleApiDataSource {
